content/canvas/test/test_type.extend.html

<!DOCTYPE HTML>
<title>Canvas test: type.extend</title>
<!-- Testing: HTMLCanvasElement methods can be added, and the new methods used by canvases -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

window.HTMLCanvasElement.prototype.getZero = function () { return 0; };
ok(canvas.getZero() === 0, "canvas.getZero() === 0");

SimpleTest.finish();

});
</script>
